chore: updating time based test (WPB-21661)
